ST-421: Refactor
